+++ src/main/java/com/nextstep/precourse/Main.java
@@ -1,4 +1,6 @@
-package baseballgame;
+package com.nextstep.precourse;
+
+import com.nextstep.precourse.baseballgame.BaseBallGame;
 
 public class Main {
 	public static void main(String[] args) {
