+++ src/test/java/com/nextstep/precourse/BallTest.java
@@ -1,4 +1,4 @@
-package baseballgame;
+package com.nextstep.precourse;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -9,6 +9,8 @@ import org.junit.jupiter.params.provider.ValueSource;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+
+import com.nextstep.precourse.computer.Ball;
 
 public class BallTest {
