+++ src/test/java/com/nextstep/precourse/UserTest.java
@@ -1,4 +1,4 @@
-package baseballgame;
+package com.nextstep.precourse;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -6,6 +6,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import com.nextstep.precourse.user.User;
+import com.nextstep.precourse.util.InputValidator;
 
 public class UserTest {
